Make workspace mandatory
We are choosing this option for now, as we have not decided how to deal
with empty workspaces yet. There are several options like defaulting to
the `ungrouped` workspace, or to the `default` workspace.

For now, let's force the clients to specify the workspace, as this is
the most future-proof approach. If in the future we need to change this,
it will not be a breaking change.